drivers: msgintr/netc: convert pointers to uintptr_t for aarch64

+++ drivers/msgintr/fsl_msgintr.h
@@ -62,7 +62,7 @@ static inline uint32_t MSGINTR_GetIntrSelectAddr(MSGINTR_Type *base, uint8_t cha
 {
     assert(channel < FSL_MSGINTR_CHANNEL_NUM);
 
-    return (uint32_t)(&base->MSI[channel].MSIIR);
+    return (uintptr_t)(&base->MSI[channel].MSIIR);
 }
 
 #if defined(__cplusplus)
